requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.12
google-genai>=1.13.0,<2.0.0
